More clean up

+++ db/migrate/20200222091451_drop_city_assoc_tables.rb
@@ -2,14 +2,14 @@
 
 class DropCityAssocTables < ActiveRecord::Migration[5.1]
   def change
-    drop_table(:feedbacks)
-    drop_table(:articles)
-    drop_table(:city_articles)
-    drop_table(:faqs)
-    drop_table(:overall_averages)
-    drop_table(:rates)
-    drop_table(:rating_caches)
-    drop_table(:articles_categories)
-    drop_table(:average_caches)
+    drop_table(:feedbacks) if table_exists?(:feedbacks)
+    drop_table(:articles) if table_exists?(:articles)
+    drop_table(:city_articles) if table_exists?(:city_articles)
+    drop_table(:faqs) if table_exists?(:faqs)
+    drop_table(:overall_averages) if table_exists?(:overall_averages)
+    drop_table(:rates) if table_exists?(:rates)
+    drop_table(:rating_caches) if table_exists?(:rating_caches)
+    drop_table(:articles_categories) if table_exists?(:articles_categories)
+    drop_table(:average_caches) if table_exists?(:average_caches)
   end
 end
